requirements.txt
Django==5.0.2
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
setuptools==79.0.1 